+++ OneBlink.SDK/models/Form.cs
@@ -322,5 +322,9 @@ namespace OneBlink.SDK.Model
         {
             get; set;
         }
+        public bool? isAIBuilderSupported
+        {
+            get; set;
+        }
     }
 }